+++ CotEditor/Sources/CEAppDelegate.h
@@ -40,3 +40,12 @@
 - (IBAction)openHelpAnchor:(id)sender;
 
 @end
+
+
+
+
+@interface CEAppDelegate (Services)
+
+- (void)openSelection:(NSPasteboard *)pboard userData:(NSString *)userData error:(NSString **)error;
+
+@end