+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Change Log
 
 - New AppleScript property `tab width` for document object.
 - Now, CotEditor script receives the absolute file path of the frontmost document as an argument if available.
+- Add “New CotEditor Document with Selection” Services menu.
 
 
 ### Additions/Changes